LayoutTests/inspector/console/js-isLikelyStackTrace-expected.txt

Test stack trace detection heuristic.


== Running test suite: WI.StackTrace.isLikelyStackTrace
-- Running test case: notStacktrace
PASS: Should NOT be a stacktrace.

-- Running test case: typeErrorWrap
FAIL: Should be a stacktrace.
    Expected: truthy
    Actual: false

-- Running test case: getAnonymousFunctionError
PASS: Should be a stacktrace.

-- Running test case: testWithNativeCallInBetween
FAIL: Should be a stacktrace.
    Expected: truthy
    Actual: false
